OpenAI_Learning/OmniBot/requirements.txt
flask
python-dotenv
openai
streamlit